add savecommnet() method

+++ artificialsw-backend/src/main/java/org/dcode/artificialswbackend/community/entity/Comment.java
@@ -1,8 +1,6 @@
 package org.dcode.artificialswbackend.community.entity;
 
 import jakarta.persistence.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
 
 import java.time.LocalDateTime;
 
@@ -11,7 +9,7 @@ import java.time.LocalDateTime;
 public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private Long questionId;
     private Long writer;
@@ -23,11 +21,11 @@ public class Comment {
 
     public Comment() {}
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
